js bugs

+++ public/javascript/delete.js
@@ -1,25 +1,22 @@
-async function deleteFormHandler(event) {
-    event.preventDefault();
+const delButtonHandler = async (event) => {
 
-    const id = window.location.toString().split('/')[
-        window.location.toString().split('/').length - 1
-    ];
-
-    const response = await fetch(`/api/posts/${id}`, {
+    if (event.target.hasAttribute('data-id')) {
+      const id = event.target.getAttribute('data-id');
+        console.log(id);
+      const response = await fetch(`/api/post/${id}`, {
         method: 'DELETE',
-        body: JSON.stringify({
-            post_id: id
-        }),
-        headers: {
-            'Content-Type': 'application/json'
-        }
-    });
-    
-    if (response.ok) {
-        document.location.replace('/dashboard');
-    } else {
-        alert(response.statusText);
+      });
+  
+      if (response.ok) {
+        document.location.replace('/');
+      } else {
+        alert('Failed to delete post');
+      }
     }
-}
 
-document.querySelector('#delete').addEventListener('click', deleteFormHandler);
+  };
+ 
+ 
+ document
+    .querySelector('#deleteBtn')
+    .addEventListener('click', delButtonHandler);
